fix: statusBar 보이기 숨기기 함수 추가
