+++ PayKids/presentation/src/main/res/layout/fragment_diary.xml
@@ -3,15 +3,16 @@
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:background="@color/gray5"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:background="@color/gray5">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@color/gray5"
-        android:padding="12dp">
+        android:paddingHorizontal="12dp"
+        android:paddingTop="32dp">
 
         <ImageButton
             android:id="@+id/ib_left"
